requirements.txt
pandas
openai
azure-search-documents
azure-core